#325: Move ITs to correct packages.
